Add exceptions in Team and TeamLeader services
Comment delete tests (failed)

ADD

+++ src/main/java/fr/enssat/leave_manager/service/impl/TeamServiceImpl.java
@@ -3,6 +3,7 @@ package fr.enssat.leave_manager.service.impl;
 import fr.enssat.leave_manager.model.TeamEntity;
 import fr.enssat.leave_manager.repository.TeamRepository;
 import fr.enssat.leave_manager.service.TeamService;
+import fr.enssat.leave_manager.service.exception.already_exists.TeamAlreadyExistsException;
 import fr.enssat.leave_manager.service.exception.not_found.TeamNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
@@ -37,16 +38,22 @@ public class TeamServiceImpl implements TeamService {
 
     @Override
     public TeamEntity addTeam(TeamEntity team) {
+        if (exists(team.getId()))
+            throw new TeamAlreadyExistsException(team);
         return repository.saveAndFlush(team);
     }
 
     @Override
     public TeamEntity editTeam(TeamEntity team) {
+        if (!exists(team.getId()))
+            throw new TeamNotFoundException(team.getId());
         return repository.saveAndFlush(team);
     }
 
     @Override
     public void deleteTeam(String id) {
+        if (!exists(id))
+            throw new TeamNotFoundException(id);
         repository.deleteById(id);
     }
 }
